Disable code select for 'this' keyword when used in property expression
Ex: this.toString()

No need to provide info on 'this' since you are editing the type in
question.  Java Editor does not provide code select on this keyword.

+++ base/org.eclipse.jdt.groovy.core/src/org/eclipse/jdt/groovy/core/util/GroovyUtils.java
@@ -113,7 +113,7 @@ public abstract class GroovyUtils {
      */
     public static int startOffset(AnnotationNode node) {
         int start = -1;
-        Long offsets = node.getNodeMetaData("source.offsets");
+        Long offsets = (Long) node.getNodeMetaData("source.offsets");
         if (offsets != null) {
             start = (int) (offsets >> 32);
         } else if (node.getEnd() > 0) {
@@ -127,7 +127,7 @@ public abstract class GroovyUtils {
      */
     public static int endOffset(AnnotationNode node) {
         int end = -1;
-        Long offsets = node.getNodeMetaData("source.offsets");
+        Long offsets = (Long) node.getNodeMetaData("source.offsets");
         if (offsets != null) {
             end = (int) (offsets & 0xFFFFFFFF);
         } else {
